+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added xExchAddressList ressource
 - Added xExchSendConnector resource
+- Added additional parameters to the MSFT_xExchImapSettings resource
 
 ## 1.30.0.0